x_font: workaround for proportional font rendering
Proportional font character offsets are stored as 32-bit
values in the file. If 'long' is not a 32-bit type,
such as 64-bit Linux systems, loading this data fails,
which causes proportional fonts to render incorrectly.

Use an intermediate 'int' array to read the values for now;
the proper solution is migrating to well-defined types, such
as uint32_t from stdio.h - which Borland C++ 3.1 lacks.