config: migrate to new config API (part 4)
This patch migrates the $F handling in MIME.CFG,
but the correct behaviour is not fully clear.

The v1.99 configuration does not contain FILEargs,
therefore the existing code copies its byte from
the NULL pointer instead of the configuration file.

Just taking a guess here is less bad, at least.