arachne: remove dotarachne, move textarea.tmp
The location for user data is now stored in 'userpath',
which is defined in all builds. Therefore, the previous
POSIX-only 'dotarachne' is no longer needed and removed.

This change also moves the location of 'textarea.tmp'
from the current directory to the userpath, avoiding
clutter and issues if the current directory is read-only.