font: fix proportional font rendering for 64-bit
The content of xg_fonadr is read from a file,
and the array type mismatch caused rendering of
proportional fonts to fail on 64-bit builds.

Declare xg_fonadr as int32_t instead of "long int"
to match the file layout in all builds. Remove the
workaround (which only worked for ASCII characters
anyway).

Note that this causes a type mismatch between LOPIF
and the Arachne core. However, LOPIF exclusively
targets DOS where both types are equivalent.